fix conflicts, cosmetic changes as per the review
