+++ src/main/java/com/autotune/analyzer/services/GenerateRecommendations.java
@@ -129,143 +129,6 @@ public class GenerateRecommendations extends HttpServlet {
         }
     }
 
-    public static void fetchMetricsBasedOnDatasource(KruizeObject kruizeObject, Timestamp interval_end_time, Timestamp interval_start_time, DataSourceInfo dataSourceInfo) throws Exception {
-        try {
-            long interval_end_time_epoc = interval_end_time.getTime() / KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC
-                    - ((long) interval_end_time.getTimezoneOffset() * KruizeConstants.TimeConv.NO_OF_SECONDS_PER_MINUTE);
-            long interval_start_time_epoc = interval_start_time.getTime() / KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC
-                    - ((long) interval_start_time.getTimezoneOffset() * KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC);
-            // Get MetricsProfile name and list of promQL to fetch
-            Map<AnalyzerConstants.MetricName, String> promQls = new HashMap<>();
-            getPromQls(promQls);
-            List<String> aggregationMethods = Arrays.asList(KruizeConstants.JSONKeys.SUM, KruizeConstants.JSONKeys.AVG,
-                    KruizeConstants.JSONKeys.MAX, KruizeConstants.JSONKeys.MIN);
-            Double measurementDurationMinutesInDouble = kruizeObject.getTrial_settings().getMeasurement_durationMinutes_inDouble();
-            List<K8sObject> kubernetes_objects = kruizeObject.getKubernetes_objects();
-            for (K8sObject k8sObject : kubernetes_objects) {
-                String namespace = k8sObject.getNamespace();
-                HashMap<String, ContainerData> containerDataMap = k8sObject.getContainerDataMap();
-                for (Map.Entry<String, ContainerData> entry : containerDataMap.entrySet()) {
-                    ContainerData containerData = entry.getValue();
-                    String containerName = containerData.getContainer_name();
-                    HashMap<Timestamp, IntervalResults> containerDataResults = new HashMap<>();
-                    IntervalResults intervalResults;
-                    HashMap<AnalyzerConstants.MetricName, MetricResults> resMap;
-                    MetricResults metricResults;
-                    MetricAggregationInfoResults metricAggregationInfoResults;
-                    for (Map.Entry<AnalyzerConstants.MetricName, String> metricEntry : promQls.entrySet()) {
-                        for (String methodName : aggregationMethods) {
-                            String promQL = null;
-                            String format = null;
-                            if (metricEntry.getKey() == AnalyzerConstants.MetricName.cpuUsage) {
-                                String secondMethodName = methodName;
-                                if (secondMethodName.equals(KruizeConstants.JSONKeys.SUM))
-                                    secondMethodName = KruizeConstants.JSONKeys.AVG;
-                                promQL = String.format(metricEntry.getValue(), methodName, secondMethodName, namespace, containerName, measurementDurationMinutesInDouble.intValue());
-                                format = KruizeConstants.JSONKeys.CORES;
-                            } else if (metricEntry.getKey() == AnalyzerConstants.MetricName.cpuThrottle) {
-                                promQL = String.format(metricEntry.getValue(), methodName, namespace, containerName, measurementDurationMinutesInDouble.intValue());
-                                format = KruizeConstants.JSONKeys.CORES;
-                            } else if (metricEntry.getKey() == AnalyzerConstants.MetricName.cpuLimit || metricEntry.getKey() == AnalyzerConstants.MetricName.cpuRequest) {
-                                promQL = String.format(metricEntry.getValue(), methodName, namespace, containerName);
-                                format = KruizeConstants.JSONKeys.CORES;
-                            } else if (metricEntry.getKey() == AnalyzerConstants.MetricName.memoryUsage || metricEntry.getKey() == AnalyzerConstants.MetricName.memoryRSS) {
-                                String secondMethodName = methodName;
-                                if (secondMethodName.equals(KruizeConstants.JSONKeys.SUM))
-                                    secondMethodName = KruizeConstants.JSONKeys.AVG;
-                                promQL = String.format(metricEntry.getValue(), methodName, secondMethodName, namespace, containerName, measurementDurationMinutesInDouble.intValue());
-                                format = KruizeConstants.JSONKeys.GIBIBYTE;
-                            } else if (metricEntry.getKey() == AnalyzerConstants.MetricName.memoryLimit || metricEntry.getKey() == AnalyzerConstants.MetricName.memoryRequest) {
-                                promQL = String.format(metricEntry.getValue(), methodName, namespace, containerName);
-                                format = KruizeConstants.JSONKeys.GIBIBYTE;
-                            }
-                            if (promQL != null) {
-                                LOGGER.info(promQL);
-                                String podMetricsUrl;
-                                try {
-                                    podMetricsUrl = String.format(KruizeConstants.DataSourceConstants.DATASOURCE_ENDPOINT_WITH_QUERY,
-                                            dataSourceInfo.getUrl(),
-                                            URLEncoder.encode(promQL, CHARACTER_ENCODING),
-                                            interval_start_time_epoc,
-                                            interval_end_time_epoc,
-                                            measurementDurationMinutesInDouble.intValue() * 60);
-                                    LOGGER.info(podMetricsUrl);
-                                    JSONObject genericJsonObject = new GenericRestApiClient(podMetricsUrl).fetchMetricsJson("get", "");
-                                    JsonObject jsonObject = new Gson().fromJson(genericJsonObject.toString(), JsonObject.class);
-                                    JsonArray resultArray = jsonObject.getAsJsonObject(KruizeConstants.JSONKeys.DATA).getAsJsonArray(KruizeConstants.DataSourceConstants.DataSourceQueryJSONKeys.RESULT);
-                                    if (null != resultArray && !resultArray.isEmpty()) {
-                                        resultArray = jsonObject.getAsJsonObject(KruizeConstants.JSONKeys.DATA).getAsJsonArray(
-                                                        KruizeConstants.DataSourceConstants.DataSourceQueryJSONKeys.RESULT).get(0)
-                                                .getAsJsonObject().getAsJsonArray(KruizeConstants.DataSourceConstants
-                                                        .DataSourceQueryJSONKeys.VALUES);
-                                        SimpleDateFormat sdf = new SimpleDateFormat(KruizeConstants.DateFormats.STANDARD_JSON_DATE_FORMAT, Locale.ROOT);
-                                        sdf.setTimeZone(TimeZone.getTimeZone(KruizeConstants.TimeUnitsExt.TimeZones.UTC));
-                                        // Find LongTerm duration  keep Start , End and Step(measurement Duration)
-                                        Timestamp sTime = interval_start_time;
-                                        for (JsonElement element : resultArray) {
-                                            JsonArray valueArray = element.getAsJsonArray();
-                                            long epochTime = valueArray.get(0).getAsLong();
-                                            double value = valueArray.get(1).getAsDouble();
-                                            String timestamp = sdf.format(new Date(epochTime * KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC));
-                                            Date date = sdf.parse(timestamp);
-                                            Timestamp eTime = new Timestamp(date.getTime());
-                                            if (containerDataResults.containsKey(eTime)) {
-                                                intervalResults = containerDataResults.get(eTime);
-                                                resMap = intervalResults.getMetricResultsMap();
-                                            } else {
-                                                intervalResults = new IntervalResults();
-                                                resMap = new HashMap<>();
-                                            }
-                                            if (resMap.containsKey(metricEntry.getKey())) {
-                                                metricResults = resMap.get(metricEntry.getKey());
-                                                metricAggregationInfoResults = metricResults.getAggregationInfoResult();
-                                            } else {
-                                                metricResults = new MetricResults();
-                                                metricAggregationInfoResults = new MetricAggregationInfoResults();
-                                            }
-                                            Method method = MetricAggregationInfoResults.class.getDeclaredMethod("set" + methodName.substring(0, 1).toUpperCase() + methodName.substring(1), Double.class);
-                                            method.invoke(metricAggregationInfoResults, value);
-                                            metricAggregationInfoResults.setFormat(format);
-                                            metricResults.setAggregationInfoResult(metricAggregationInfoResults);
-                                            metricResults.setName(String.valueOf(metricEntry.getKey()));
-                                            metricResults.setFormat(format);
-                                            resMap.put(metricEntry.getKey(), metricResults);
-                                            intervalResults.setMetricResultsMap(resMap);
-                                            intervalResults.setIntervalStartTime(sTime);  //Todo this will change
-                                            intervalResults.setIntervalEndTime(eTime);
-                                            intervalResults.setDurationInMinutes((double) ((eTime.getTime() - sTime.getTime())
-                                                    / ((long) KruizeConstants.TimeConv.NO_OF_SECONDS_PER_MINUTE
-                                                    * KruizeConstants.TimeConv.NO_OF_MSECS_IN_SEC)));
-                                            containerDataResults.put(eTime, intervalResults);
-                                            sTime = eTime;
-                                        }
-                                    }
-                                } catch (Exception e) {
-                                    throw new RuntimeException(e);
-                                }
-                            }
-                        }
-                    }
-                    containerData.setResults(containerDataResults);
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new Exception("Exception occurred while fetching metrics from the datasource: "+e.getMessage());
-        }
-    }
-
-    private static void getPromQls(Map<AnalyzerConstants.MetricName, String> promQls) {
-        promQls.put(AnalyzerConstants.MetricName.cpuUsage, PromQLDataSourceQueries.CPU_USAGE);
-        promQls.put(AnalyzerConstants.MetricName.cpuThrottle, PromQLDataSourceQueries.CPU_THROTTLE);
-        promQls.put(AnalyzerConstants.MetricName.cpuLimit,PromQLDataSourceQueries.CPU_LIMIT);
-        promQls.put(AnalyzerConstants.MetricName.cpuRequest,PromQLDataSourceQueries.CPU_REQUEST);
-        promQls.put(AnalyzerConstants.MetricName.memoryUsage, PromQLDataSourceQueries.MEMORY_USAGE);
-        promQls.put(AnalyzerConstants.MetricName.memoryRSS, PromQLDataSourceQueries.MEMORY_RSS);
-        promQls.put(AnalyzerConstants.MetricName.memoryLimit, PromQLDataSourceQueries.MEMORY_LIMIT);
-        promQls.put(AnalyzerConstants.MetricName.memoryRequest, PromQLDataSourceQueries.MEMORY_REQUEST);
-    }
-
     private void sendSuccessResponse(HttpServletResponse response, KruizeObject ko, Timestamp interval_end_time) throws IOException {
         LOGGER.debug("sendSuccessResponse");
         response.setContentType(JSON_CONTENT_TYPE);
